requirements.txt
psutil
python-nmap
scapy
tabulate
